Converted int coords of game objects to double coords. (We're using Path2D now instead of Polygon to make the shipBody) Feels a bit smoother now that our coords and speed factors are all doubles. Looks smoother too (Could just be me)
Also shooting works if you press space. Gotta find a way to influence fire rate (I'm thinking we set a timer every time space appears in the input list) and also we finetune optimal velocity for bullets. 

No bullet collisions yet. (To be fair, we have nothing for them to collide with yet lol)

+++ src/asteroids/components/GameComponent.java
@@ -6,29 +6,29 @@ import java.awt.*;
 
 public abstract class GameComponent {
 
-    protected int x;
-    protected int y;
+    protected double x;
+    protected double y;
     protected Color color;
 
-    public GameComponent(int x, int y, Color color) {
+    public GameComponent(double x, double y, Color color) {
         this.x = x;
         this.y = y;
         this.color = color;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
